+++ api/V0ComposePush.go
@@ -14,7 +14,7 @@ import (
 
 // V0ComposePush will read and interpret the docker-compose.yml
 // example:
-// curl -X GET http://user:password@127.0.0.1:10000/v0/compose --data-binary @docker-compose.yml
+// curl -X GET http://user:password@127.0.0.1:10000/api/compose/v0 --data-binary @docker-compose.yml
 func V0ComposePush(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	auth := CheckAuth(r, w)