+++ api/V0ComposeRestartService.go
@@ -12,7 +12,7 @@ import (
 
 // V0ComposeRestartService will restart a service from a specific project
 // example:
-// curl -X PUT http://user:password@127.0.0.1:10000/v0/compose/{projectname}/{servicename}/restart
+// curl -X PUT http://user:password@127.0.0.1:10000/api/compose/v0/{projectname}/{servicename}/restart
 func V0ComposeRestartService(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	auth := CheckAuth(r, w)